+++ .prettierrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  singleQuote: true,
-  semi: false,
-  printWidth: 110,
-  trailingComma: 'es5',
-  arrowParens: 'avoid',
-}