+++ config.d.ts
@@ -1,0 +1,5 @@
+declare const _default: {
+    baseURL: string;
+    timeout: number;
+};
+export default _default;